refactor: separate Puck layout into own file
